unit_test: move lattice constant test into unit_test directory
Move the lattice constant test into the unit_test directory and rewrite
it to pytest style.
